@tbg 2022-12-22
Thank you! I rebased your PR, there were merge conflicts due to #8.

Re-ran the benchmarks produced different numbers now, but the reduction we're targeting is still there.

```
$ benchdiff --old HEAD~1 -r BenchmarkRawNode -c 5 .
test binaries already exist for '65a0bf3'; skipping build
test binaries already exist for '228ee70'; skipping build

  pkg=1/1 iter=5/5 go.etcd.io/raft/v3 /

name                     old time/op        new time/op        delta
RawNode/single-voter-10         918ns ± 1%         900ns ± 1%   -1.93%  (p=0.032 n=5+5)
RawNode/two-voters-10          1.37µs ± 1%        1.35µs ± 1%   -1.71%  (p=0.008 n=5+5)

name                     old firstIndex/op  new firstIndex/op  delta
RawNode/single-voter-10          1.00 ± 0%          1.00 ± 0%     ~     (all equal)
RawNode/two-voters-10            3.00 ± 0%          3.00 ± 0%     ~     (all equal)

name                     old lastIndex/op   new lastIndex/op   delta
RawNode/single-voter-10          2.00 ± 0%          2.00 ± 0%     ~     (all equal)
RawNode/two-voters-10            2.00 ± 0%          2.00 ± 0%     ~     (all equal)

name                     old ready/op       new ready/op       delta
RawNode/single-voter-10          2.00 ± 0%          2.00 ± 0%     ~     (all equal)
RawNode/two-voters-10            2.00 ± 0%          2.00 ± 0%     ~     (all equal)

name                     old term/op        new term/op        delta
RawNode/single-voter-10          0.00 ± 0%          0.00 ± 0%  -12.50%  (p=0.008 n=5+5)
RawNode/two-voters-10            1.00 ± 0%          1.00 ± 0%     ~     (all equal)

name                     old alloc/op       new alloc/op       delta
RawNode/two-voters-10            883B ± 1%          850B ± 1%   -3.72%  (p=0.008 n=5+5)
RawNode/single-voter-10          533B ± 0%          549B ± 2%   +2.89%  (p=0.008 n=5+5)

name                     old allocs/op      new allocs/op      delta
RawNode/single-voter-10          6.00 ± 0%          4.00 ± 0%  -33.33%  (p=0.008 n=5+5)
RawNode/two-voters-10            8.00 ± 0%          6.00 ± 0%  -25.00%  (p=0.008 n=5+5)
```

Looks like #8 (or something else) gave us an extra allocation.

Not this PR's problem, though.